requirements/base.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
idna==2.10
multidict==5.1.0
pytz==2021.1
requests==2.25.1
slackclient==2.9.3
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
cryptography==3.4.7
python-decouple==3.4